+++ examples/egt.py
@@ -1,0 +1,122 @@
+import os.path as osp
+
+import torch
+import torch.nn.functional as F
+from ogb.graphproppred import Evaluator
+from ogb.graphproppred import PygGraphPropPredDataset as OGBG
+from ogb.graphproppred.mol_encoder import AtomEncoder, BondEncoder
+from torch.nn import BatchNorm1d, Linear, ReLU, Sequential
+from torch.optim.lr_scheduler import ReduceLROnPlateau
+
+import torch_geometric.transforms as T
+from torch_geometric.loader import DataLoader
+from torch_geometric.nn import EGTConv, global_mean_pool
+from torch_geometric.typing import WITH_TORCH_SPARSE
+from torch_geometric.utils import to_dense_adj, to_edge_index
+
+if not WITH_TORCH_SPARSE:
+    quit("This example requires 'torch-sparse'")
+
+path = osp.join(osp.dirname(osp.realpath(__file__)), '..', 'data', 'OGB')
+dataset = OGBG('ogbg-molhiv', path, pre_transform=T.ToSparseTensor())
+evaluator = Evaluator('ogbg-molhiv')
+
+split_idx = dataset.get_idx_split()
+train_dataset = dataset[split_idx['train']]
+val_dataset = dataset[split_idx['valid']]
+test_dataset = dataset[split_idx['test']]
+
+train_loader = DataLoader(train_dataset, batch_size=32, num_workers=4,
+                          shuffle=True)
+val_loader = DataLoader(val_dataset, batch_size=256)
+test_loader = DataLoader(test_dataset, batch_size=256)
+
+
+class Net(torch.nn.Module):
+    def __init__(self, hidden_channels, num_layers):
+        super().__init__()
+
+        self.atom_encoder = AtomEncoder(hidden_channels)
+        self.bond_encoder = BondEncoder(hidden_channels)
+
+        self.convs = torch.nn.ModuleList()
+        for _ in range(num_layers):
+            self.convs.append(
+                EGTConv(channels=hidden_channels, edge_dim=hidden_channels,
+                        edge_update=False, heads=4, attn_dropout=0.3,
+                        num_virtual_nodes=0))
+
+        self.mlp = Sequential(
+            Linear(hidden_channels, hidden_channels // 2, bias=False),
+            BatchNorm1d(hidden_channels // 2),
+            ReLU(inplace=True),
+            Linear(hidden_channels // 2, hidden_channels // 4, bias=False),
+            BatchNorm1d(hidden_channels // 4),
+            ReLU(inplace=True),
+            Linear(hidden_channels // 4, 1),
+        )
+
+    def forward(self, x, adj_t, edge_attr, batch):
+        x = self.atom_encoder(x)
+        edge_attr = self.bond_encoder(edge_attr)
+        edge_index, _ = to_edge_index(adj_t)
+        edge_attr = to_dense_adj(edge_index, batch, edge_attr)
+
+        for conv in self.convs:
+            x = conv(x, edge_attr, batch)
+
+        x = global_mean_pool(x, batch)
+
+        return self.mlp(x)
+
+
+device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+model = Net(hidden_channels=128, num_layers=8).to(device)
+
+optimizer = torch.optim.Adam(model.parameters(), lr=1e-4)
+scheduler = ReduceLROnPlateau(optimizer, mode='max', factor=0.5, patience=20,
+                              min_lr=1e-5)
+
+
+def train():
+    model.train()
+
+    total_loss = total_examples = 0
+    for data in train_loader:
+        data = data.to(device)
+        optimizer.zero_grad()
+
+        out = model(data.x, data.adj_t, data.edge_attr, data.batch)
+        loss = F.binary_cross_entropy_with_logits(out, data.y.to(torch.float))
+        loss.backward()
+        optimizer.step()
+
+        total_loss += float(loss) * data.num_graphs
+        total_examples += data.num_graphs
+
+    return total_loss / total_examples
+
+
+@torch.no_grad()
+def evaluate(loader):
+    model.eval()
+
+    y_pred, y_true = [], []
+    for data in loader:
+        data = data.to(device)
+        pred = model(data.x, data.adj_t, data.edge_attr, data.batch)
+        y_pred.append(pred.cpu())
+        y_true.append(data.y.cpu())
+
+    y_true = torch.cat(y_true, dim=0)
+    y_pred = torch.cat(y_pred, dim=0)
+    return evaluator.eval({'y_true': y_true, 'y_pred': y_pred})['rocauc']
+
+
+for epoch in range(1, 31):
+    loss = train()
+    val_rocauc = evaluate(val_loader)
+    test_rocauc = evaluate(test_loader)
+    scheduler.step(val_rocauc)
+    print(f'Epoch: {epoch:02d}, Loss: {loss:.4f}, Val: {val_rocauc:.4f}, '
+          f'Test: {test_rocauc:.4f}')